requirements.txt
amqp==2.3.2
atomicwrites==1.2.1
attrs==18.2.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
djangorestframework==3.9.0
envparse==0.2.0
flake8==3.6.0
idna==2.7
kombu==4.2.1
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.8.0
psycopg2==2.7.6.1
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
pytest==4.0.1
pytest-django==3.4.4
pytz==2018.7
redis==3.0.1
requests==2.20.1
six==1.11.0
urllib3==1.24.1
vine==1.1.4
